# order-service/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Order Service API
  description: API для управления заказами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

paths:
  /orders/order/{orderId}:
    get:
      tags:
        - Orders
      summary: Получить заказ по ID
      description: Возвращает информацию о конкретном заказе
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID заказа
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Неверные параметры запроса
        '404':
          description: Заказ не найден

  /orders/all/{userId}:
    get:
      tags:
        - Orders
      summary: Получить все заказы пользователя
      description: Возвращает список всех заказов для указанного пользователя
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Неверные параметры запроса

  /orders/{orderId}:
    delete:
      tags:
        - Orders
      summary: Удалить заказ
      description: Удаляет заказ по указанному ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID заказа для удаления
      responses:
        '200':
          description: Заказ успешно удален
        '400':
          description: Неверные параметры запроса

  /orders/{userId}:
    post:
      tags:
        - Orders
      summary: Создать новый заказ
      description: Создает новый заказ для указанного пользователя
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDTO'
      responses:
        '200':
          description: Успешное создание заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Неверные параметры запроса

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор заказа
        name:
          type: string
          description: Название заказа
        count:
          type: integer
          format: int64
          description: Количество товаров в заказе
        userId:
          type: integer
          format: int64
          description: ID пользователя
        creationDate:
          type: string
          format: date-time
          description: Дата создания заказа
      required:
        - name
        - count
        - userId

    OrderDTO:
      type: object
      properties:
        name:
          type: string
          description: Название заказа
        count:
          type: integer
          format: int64
          description: Количество товаров в заказе
      required:
        - name
        - count

  responses:
    BadRequest:
      description: Неверные параметры запроса
    NotFound:
      description: Ресурс не найден